chore(website): G-CQ07 S2-Import-Zyklen entkoppeln (5→0) [T001179]
Loest alle 5 zirkulaeren Abhaengigkeiten im website-Modulgraph via
gezielte Modul-Extraktion. Superset von #2114 (T001172), das
Zyklus #1 nur halb brach (tickets-schema.ts importierte weiterhin
pool aus website-db.ts → neuer Zyklus tickets-schema↔website-db).

### Refactor

1) website/src/lib/db-pool.ts (neu) — pg.Pool + ensureSchemaOnce-Cache
   extrahiert. Leaf-Modul ohne Projekt-Importe. website-db.ts re-exportiert
   pool/ensureSchemaOnce/platformPool/__resetSchemaInitCacheForTests fuer
   externe Aufrufer.

2) tickets-schema.ts, tickets/transition.ts, tickets/reporter-link.ts:
   Import-Ziel von './website-db' → './db-pool' (Cycle-Bruch fuer
   S2:website:src/lib/{tickets-schema,website-db,transition,reporter-link}).

3) website/src/lib/invoice-types.ts (neu) — Invoice + InvoiceLine
   extrahiert. Leaf-Modul. native-billing.ts re-exportiert fuer
   externe Aufrufer (api/admin/billing/create-invoice,
   finalize-from-prepayment).

4) invoice-pdf.ts: Import-Ziel type Invoice von './native-billing'
   → './invoice-types'. Funktion embedFacturX entfernt.

5) pdf-a3-embed.ts: embedFacturX hinzugefuegt (war in invoice-pdf.ts
   mit dynamic import('einvoice/sidecar-client')). Sidecar-Imports
   sind jetzt statisch (Leaf-Modul sidecar-client).

6) native-billing.ts: dynamic import('./invoice-pdf') →
   dynamic import('./pdf-a3-embed') (Cycle-Bruch fuer
   S2:website:src/lib/{invoice-pdf,native-billing}).

7) tickets-db.featureflag.test.ts: vi.mock('./website-db') →
   vi.mock('./db-pool') (Mock-Ziel entsprechend verschoben).

8) docs/code-quality/baseline.json: 5 S2-FIXED-Eintraege via
   task quality:baseline:refresh entfernt.

### Effekt

- G-CQ07: 5 → 0 (Ziel <=0) — erstmalig gruen.
- G-RH01: 33 → 28 (Ziel <=30) — erstmalig gruen!
- Vitest: 1510/1510 passed (0 regressions).
- madge --circular: 0 cycles (vorher 4+1).
- quality:check: 28 current / 28 baselined / 0 blocking.

### Audit-Ticket

T001179 (type=task, status=done — Chore, kein Plan-Handoff).
Superseded: T001178 (PR #2117, DIRTY nach #2114-Race).
