+++ website/src/lib/tickets-schema.ts
@@ -2,7 +2,7 @@
 // Schema-Initialisierung + Helper für das `tickets`-PostgreSQL-Schema.
 // Ausgelagert aus tickets-db.ts (G-CQ07, Zyklus #1), um den statischen
 // Import-Zyklus zwischen tickets-db.ts und website-db.ts aufzubrechen.
-import { pool, ensureSchemaOnce } from './website-db';
+import { pool, ensureSchemaOnce } from './db-pool';
 import { MixedEmbeddingModelError } from './knowledge-db';
 import { initProviderConfigSchema } from './schema/provider-config-schema';
 import { applyTicketsCoreSchema } from './tickets/tables/tickets';
